refac: reorganize security package structure and update token service implementation
