+++ src/main/java/collabhubbr/users/service/TokenService.java
@@ -1,0 +1,8 @@
+package collabhubbr.users.service;
+
+import collabhubbr.users.models.UserEntity;
+
+public interface TokenService {
+    String generateToken(UserEntity user);
+    String validateToken(String token);
+}
